db/pyproject.toml
[project]
name = "db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.32",
]
